--- index.js
@@ -1,4 +1,0 @@
-module.exports = require('./ActionSheet');
-
-module.exports.ActionSheetProvider = require('./ActionSheetProvider').default;
-module.exports.connectActionSheet = require('./connectActionSheet').default;
