+++ src/ActionSheet/index.ios.tsx
@@ -1,22 +1,25 @@
-'use strict';
-
-import React from 'react';
+import * as React from 'react';
 import { ActionSheetIOS, View } from 'react-native';
+import { ActionSheetIOSOptions } from '../types';
 
-export default class ActionSheet extends React.Component {
+type Props = {
+  readonly children: React.ReactNode;
+};
+
+type onSelect = (buttonIndex: number) => void;
+
+export default class ActionSheet extends React.Component<Props> {
   render() {
     return <View style={{ flex: 1 }}>{React.Children.only(this.props.children)}</View>;
   }
 
-  showActionSheetWithOptions(options, onSelect) {
-    const iosOptions = {
+  showActionSheetWithOptions(options: ActionSheetIOSOptions, onSelect: onSelect) {
+    const iosOptions: ActionSheetIOSOptions = {
       ...options,
       // A null title or message on iOS causes a crash
       title: options.title || undefined,
       message: options.message || undefined,
-      icons: undefined,
     };
-
     ActionSheetIOS.showActionSheetWithOptions(iosOptions, onSelect);
   }
 }
