+++ src/index.ts
@@ -1,0 +1,4 @@
+export { default } from './ActionSheet';
+export { default as ActionSheetProvider } from './ActionSheetProvider';
+export { default as connectActionSheet } from './connectActionSheet';
+export * from './types';
