+++ tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "expo-module-scripts/tsconfig.base",
+  "compilerOptions": {
+    "outDir": "./build",
+  },
+  "include": ["./src"]
+}
